requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.3
transformers==4.34.0
torch==2.0.1
scikit-learn==1.3.0
gensim==4.3.1
pybtex==0.24.0
lxml==4.9.3

Flask==2.3.2
feedparser==6.0.10
biopython==1.80
vaderSentiment==3.3.2
